# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/tool/getMessageJson": {
            "post": {
                "tags": [
                    "工具"
                ],
                "summary": "获取消息json格式",
                "parameters": [
                    {
                        "description": "消息",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "tags": [
                    "用户操作"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户手机",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "userName",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/user/delete": {
            "post": {
                "tags": [
                    "用户操作"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/user/exist": {
            "get": {
                "tags": [
                    "用户操作"
                ],
                "summary": "判断是否存在用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户手机",
                        "name": "phone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/user/find": {
            "get": {
                "tags": [
                    "用户操作"
                ],
                "summary": "获取用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "被查询用户Id",
                        "name": "queriedId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "用户操作"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/user/update": {
            "post": {
                "tags": [
                    "用户操作"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户json信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserBasic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.GroupBasic": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "icon": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserBasic"
                    }
                },
                "name": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "content": {},
                "contentType": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "msgType": {
                    "type": "integer"
                },
                "receiverId": {
                    "type": "string"
                },
                "senderId": {
                    "type": "string"
                }
            }
        },
        "models.UserBasic": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GroupBasic"
                    }
                },
                "homeAddress": {
                    "type": "string"
                },
                "icon": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "loginTime": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "server.RespJson": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "data": {},
                "msg": {}
            }
        }
    }
}